Refactor: Login Retrofit Service
