+++ app/src/main/java/com/example/geeksasaeng/Login/Retrofit/LoginDataService.kt
@@ -13,6 +13,7 @@ class LoginDataService() {
 
     private lateinit var loginView: LoginView
     private lateinit var autoLoginView: AutoLoginView
+    private lateinit var socialLoginView: SocialLoginView
 
     fun setLoginView(loginView: LoginView) {
         this.loginView = loginView
@@ -20,6 +21,10 @@ class LoginDataService() {
 
     fun setAutoLoginView(autoLoginView: AutoLoginView){
         this.autoLoginView = autoLoginView
+    }
+
+    fun setSocialLoginView(socialLoginView: SocialLoginView) {
+        this.socialLoginView = socialLoginView
     }
 
     fun login(user: Login) {
@@ -62,6 +67,27 @@ class LoginDataService() {
             override fun onFailure(call: Call<AutoLoginResponse>, t: Throwable) {
                 //실패처리
                 Log.d("LOGIN-RESPONSE", "SignupDataService-onFailure : SignupFailed-자동로그인", t)
+            }
+        })
+    }
+
+    fun socialLogin(accessToken: SocialLogin){
+        val socialLoginService = NetworkModule.getInstance()?.create(LoginRetrofitInterfaces::class.java)
+        socialLoginService?.naverLogin(accessToken)?.enqueue(object: Callback<SocialLoginResponse>{
+            override fun onResponse(call: Call<SocialLoginResponse>, response: Response<SocialLoginResponse>) {
+                if (response.isSuccessful && response.code() == 200) {
+                    val socialLoginResponse: SocialLoginResponse = response.body()!!
+                    when (socialLoginResponse.code) {
+                        1000 -> socialLoginView.onSocialLoginSuccess(socialLoginResponse.code, socialLoginResponse.result!!) // 로그인
+                        2807 -> socialLoginView.onSocialLoginToRegister(socialLoginResponse.message, accessToken.accessToken)
+                        4000 -> Log.d("LOGIN", "서버 오류")
+                        else -> socialLoginView.onLoginFailure(socialLoginResponse.message)
+                    }
+                }
+            }
+            override fun onFailure(call: Call<SocialLoginResponse>, t: Throwable) {
+                //실패처리
+                Log.d("LOGIN-RESPONSE", "SignupDataService-onFailure : SignupFailed", t)
             }
         })
     }
