+++ app/src/main/java/com/example/geeksasaeng/Login/Retrofit/LoginRetrofitInterfaces.kt
@@ -15,4 +15,8 @@ interface LoginRetrofitInterfaces {
     fun autoLogin(
         @Header("Authorization") jwt: String?
         ): Call<AutoLoginResponse>
+
+    // 네이버 로그인, accessToken 전달
+    @POST("/login/social")
+    fun naverLogin(@Body accessToken: SocialLogin): Call<SocialLoginResponse>
 }
