+++ app/src/main/java/com/example/geeksasaeng/Login/Retrofit/LoginView.kt
@@ -4,3 +4,8 @@ interface LoginView {
     fun onLoginSuccess(code : Int, result : LoginResult)
     fun onLoginFailure(message : String)
 }
+
+interface SocialLoginView {
+    fun onSocialLoginSuccess(code : Int, result : SocialLoginResult)
+    fun onSocialLoginToRegister(message: String, accessToken: String)
+}
